Move pkg/utils/test to pkg/transform/internal/test
