+++ pkg/transform/oauth/keystone_test.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
+	cpmatest "github.com/fusor/cpma/pkg/transform/internal/test"
 	"github.com/fusor/cpma/pkg/transform/oauth"
-	cpmatest "github.com/fusor/cpma/pkg/utils/test"
 	configv1 "github.com/openshift/api/config/v1"
 	legacyconfigv1 "github.com/openshift/api/legacyconfig/v1"
 	"github.com/stretchr/testify/assert"
